Add Plan Detail Modal to erk dash (#1966)
## Summary

Update plan for issue #1965 (rerun)

---

To checkout this PR in a fresh worktree and environment locally, run:

```
erk pr checkout 1966
```

If using Graphite, the following steps are recommended to track and link the
branch to Graphite. The `erk` queue uses vanilla git in the remote worker.

```
gt track
gt squash && gt submit -f
```